+++ src/Entities/Content.php
@@ -13,9 +13,8 @@ declare(strict_types=1);
 namespace ElGigi\HarParser\Entities;
 
 use ElGigi\HarParser\Exception\InvalidArgumentException;
-use JsonSerializable;
 
-class Content implements JsonSerializable
+class Content implements EntityInterface
 {
     public function __construct(
         protected int $size,