+++ src/Entities/EntityInterface.php
@@ -1,0 +1,25 @@
+<?php
+/*
+ * @license   https://opensource.org/licenses/MIT MIT License
+ * @copyright 2022 Ronan GIRON
+ * @author    Ronan GIRON <https://github.com/ElGigi>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code, to the root.
+ */
+
+declare(strict_types=1);
+
+namespace ElGigi\HarParser\Entities;
+
+use JsonSerializable;
+
+interface EntityInterface extends JsonSerializable
+{
+    /**
+     * Get array copy.
+     *
+     * @return array
+     */
+    public function getArrayCopy(): array;
+}